+++ src/Totem.Timeline/Runtime/FlowScope.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Totem.Runtime;
+using Totem.Threading;
 using Totem.Timeline.Area;
 
 namespace Totem.Timeline.Runtime
